pcvgen: First pass on wrapper script to pcvgen_xe and ckgen
pcvgen_xe is primarily used as a wrapper to xedit but with 'pcv'
specifics involved.

ckgen is primarily used as a wrapper for setting up a temporary
CK working environment based off the parent directory and either
generating data or manually running ck.

pcvgen ties both together so as to automate the generation of 'pcv'
related data.